refactor(auth): auth_jwt を leaf crate alc-auth-jwt に切り出し (Refs #410)
crates/alc-core/src/auth_jwt.rs のロジックを薄い leaf crate
crates/alc-auth-jwt に移設。deps は jsonwebtoken / serde / uuid / chrono /
sha2 のみ (sqlx / reqwest / ring / ts-rs / axum を持ち込まない)。

設計判断:
- User 型: leaf crate は alc-core::models::User に依存できないため、
  create_access_token は新設の AccessTokenInput (sub/email/name/tenant_id/role)
  を取る。alc-core 側の auth_jwt module に &User を受ける互換 wrapper を残し、
  呼び出し側 (crates/alc-auth 等 13 crate) の import は無変更で通る。
- ts-rs: AppClaims は serde derive のみで ts-rs derive を持たないため leaf crate
  に ts-rs を入れる必要なし (元から #[ts(...)] 不使用)。
- 再 export 方式: alc-core::auth_jwt は `pub use alc_auth_jwt::{...};` で
  JwtSecret / AppClaims / InternalClaims / current_env_label / verify_* /
  create_internal_token / refresh token hash 群を再 export。module path 互換を
  保つので alc_core::auth_jwt::... の既存 import は無変更。
- coverage_100.toml: auth_jwt の unit test 群は leaf crate に同伴移設。新規
  crates/alc-auth-jwt/src/lib.rs を mock 型で登録。alc-core 側 shim の &User
  wrapper にも専用テストを追加し 100% を維持。
