+++ crates/alc-auth-jwt/src/test_macros.rs
@@ -1,0 +1,33 @@
+#[cfg(coverage)]
+#[macro_export]
+macro_rules! test_case {
+    ($desc:expr, $body:expr) => {
+        $body
+    };
+}
+
+#[cfg(not(coverage))]
+#[macro_export]
+macro_rules! test_case {
+    ($desc:expr, $body:expr) => {{
+        print!("  ✅ {} ... ", $desc);
+        std::io::Write::flush(&mut std::io::stdout()).ok();
+        let val = $body;
+        println!("OK");
+        val
+    }};
+}
+
+#[cfg(coverage)]
+#[macro_export]
+macro_rules! test_group {
+    ($name:expr) => {};
+}
+
+#[cfg(not(coverage))]
+#[macro_export]
+macro_rules! test_group {
+    ($name:expr) => {
+        println!("\n📋 {}", $name);
+    };
+}
